+++ backend/internal/model/category.go
@@ -2,7 +2,7 @@ package model
 
 // Categories 分类模型
 type Categories struct {
-	ID          uint32 `gorm:"primaryKey;autoIncrement" json:"id"`
+	ID          uint64 `gorm:"primaryKey;autoIncrement" json:"id"`
 	Name        string `gorm:"size:50;not null" json:"name"`
 	Sort        int    `gorm:"default:0" json:"sort"`       // 对应图片中的 sort
 	Description string `gorm:"size:255" json:"description"` // 对应图片中的 description
